\name{autocorr_mat}
\alias{autocorr_mat}
\title{Returns a p-dimensional matrix with an autocorrelation (autoregressive)
structure. The value of rho determines the amount of autocorrelation.}
\usage{
  autocorr_mat(p = 100, rho = 0.9)
}
\arguments{
  \item{p}{the dimension of the matrix}

  \item{rho}{the autocorrelation value}
}
\description{
  Returns a p-dimensional matrix with an autocorrelation
  (autoregressive) structure. The value of rho determines
  the amount of autocorrelation.
}

