\docType{package}
\name{classify}
\alias{classify}
\alias{classify-package}
\title{classify}
\description{
  classify is a framework to streamline the evaluation of
  supervised classification methods and seeks to determine
  the best classifiers on a variety of simulated and
  benchmark data sets with a collection of benchmark
  metrics.
}

