\name{cov_autocorr}
\alias{cov_autocorr}
\title{Constructs a p-dimensional covariance matrix with an autocorrelation
(autoregressive) structure.}
\usage{
  cov_autocorr(p = 100, rho = 0.9)
}
\arguments{
  \item{p}{the dimension of the matrix}

  \item{rho}{the autocorrelation value}
}
\description{
  The value of rho determines the amount of
  autocorrelation.
}

