\name{generate_duin}
\alias{generate_duin}
\title{Generate mixture (non-Gaussian) data in the format described in Skurichina and Duin (1998).}
\usage{
  generate_duin(n = 100, p = 10, seed = NULL)
}
\arguments{
  \item{n}{sample size of population}

  \item{p}{dimension of the generated data}

  \item{seed}{the random number seed of the data.}
}
\description{
  Generate mixture (non-Gaussian) data in the format
  described in Skurichina and Duin (1998).
}

